Fixed overflow bug in LoopingAlphabetStream, where a negative number ended up being used in System.arraycopy.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@543440 13f79535-47bb-0310-9956-ffa450edef68

+++ java/testing/org/apache/derbyTesting/functionTests/util/streams/LoopingAlphabetStream.java
@@ -103,7 +103,7 @@ public class LoopingAlphabetStream
             return -1;
         }
         // We can only read as many bytes as there are in the stream.
-        int nonBlankLength = Math.min((int)remainingNonBlanks, length);
+        int nonBlankLength = (int)Math.min(remainingNonBlanks, (long)length);
         fillBuffer(nonBlankLength);
         int read = 0;
         // Find position of next letter in the buffer.
